+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add More Fields On Designated Screens (display_title, credit_url, cta_text, cta_url, etc.)
 * Uploading a new pdf to a content block does not auto select or show up
 * Unselected images from Batch Edit once User clicks update.
+* Add YouTube and JW Player video fields
 
 
 ## v0.3.4